wip, add service watcher as well
