+++ BaragonService/src/main/java/com/hubspot/baragon/service/listeners/KubernetesWatchListener.java
@@ -5,31 +5,37 @@ import org.slf4j.LoggerFactory;
 
 import com.google.inject.Inject;
 import com.hubspot.baragon.config.KubernetesConfiguration;
-import com.hubspot.baragon.kubernetes.KubernetesWatcher;
+import com.hubspot.baragon.kubernetes.KubernetesEndpointsWatcher;
+import com.hubspot.baragon.kubernetes.KubernetesServiceWatcher;
 
 public class KubernetesWatchListener extends AbstractLatchListener {
   private static final Logger LOG = LoggerFactory.getLogger(ElbSyncWorkerListener.class);
 
-  private final KubernetesWatcher kubernetesWatcher;
+  private final KubernetesEndpointsWatcher kubernetesEndpointsWatcher;
+  private final KubernetesServiceWatcher kubernetesServiceWatcher;
   private final KubernetesConfiguration kubernetesConfiguration;
 
   @Inject
-  public KubernetesWatchListener(KubernetesWatcher kubernetesWatcher,
+  public KubernetesWatchListener(KubernetesEndpointsWatcher kubernetesEndpointsWatcher,
+                                 KubernetesServiceWatcher kubernetesServiceWatcher,
                                  KubernetesConfiguration kubernetesConfiguration) {
-    this.kubernetesWatcher = kubernetesWatcher;
+    this.kubernetesEndpointsWatcher = kubernetesEndpointsWatcher;
+    this.kubernetesServiceWatcher = kubernetesServiceWatcher;
     this.kubernetesConfiguration = kubernetesConfiguration;
   }
 
   @Override
   public void isLeader() {
-    LOG.info("We are the leader! Starting KubernetesWatcher...");
-    kubernetesWatcher.createWatch(true);
+    LOG.info("We are the leader! Starting KubernetesWatchers...");
+    kubernetesServiceWatcher.createWatch(true);
+    kubernetesEndpointsWatcher.createWatch(true);
 
   }
 
   @Override
   public void notLeader() {
-    kubernetesWatcher.close();
+    kubernetesServiceWatcher.close();
+    kubernetesEndpointsWatcher.close();
   }
 
   @Override
